feature : add base mappers (#4)
* feat: improve struct of pom

* feat: update dto type for match with entity

* feat: add rental mapper

* feat: add user mapper

+++ src/main/java/com/openclassrooms/rental/dto/request/MessageRequest.java
@@ -7,6 +7,6 @@ import lombok.Setter;
 @Setter
 public class MessageRequest {
     private String message;
-    private Long user_id;
-    private Long rental_id;
+    private Integer user_id;
+    private Integer rental_id;
 }
