+++ src/main/java/com/openclassrooms/rental/repository/UserRepository.java
@@ -3,4 +3,4 @@ package com.openclassrooms.rental.repository;
 import com.openclassrooms.rental.model.UserEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface UserRepository extends JpaRepository<UserEntity, Long> {}
+public interface UserRepository extends JpaRepository<UserEntity, Integer> {}
